requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
discord-protos==0.0.2
discord.py==2.5.2
discord.py-self==2.0.1
Flask==3.1.0
Flask-Session==0.8.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
multidict==6.2.0
oauthlib==3.2.2
outcome==1.3.0.post0
propcache==0.3.1
protobuf==6.30.2
py-cord==2.6.1
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.31.0
setuptools==78.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
yarl==1.18.3
python-dotenv

